+++ public/index.html
@@ -5,7 +5,7 @@
       <title>Twitch: Video Module</title>
     </head>
     <body>
-      <div id="videomodule"></div>
+      <div id="video-module"></div>
       <script type="text/javascript" src="app.js"></script>
     </body>
   </html>
